requirements.txt
openai-whisper
moviepy
typer
pdfminer.six
